Implement Dynamic Pandoc
All Pandoc elements are made dynamic.
This allows for incremental updates;
The only way this used to be able was to use dyn on the top level
pandocEl widget.
With these changes only the parts of the pandoc document that are
different are updated.

The implementation uses DSum and factorDyn. The entire block is only
redrawn if the block type changes.